pyproject.toml
[tool.poetry]
name = "nijisanji-twitter-organize"
version = "0.1.0"
description = ""
authors = ["iron316 <shunta316iron@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.1"
pandas = "^1.3.4"
PyYAML = "^6.0"
tweepy = "^4.3.0"
tqdm = "^4.62.3"
black = "^21.10b0"
beautifulsoup4 = "^4.10.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
